+++ postcss.config.js
@@ -20,7 +20,8 @@ module.exports = {
         whitelistPatterns: [
           /-(leave|enter|appear)(|-(to|from|active))$/,
           /^(?!(|.*?:)cursor-move).+-move$/,
-          /^router-link(|-exact)-active$/
+          /^router-link(|-exact)-active$/,
+          /vt-*/
         ]
       }),
     require("autoprefixer")()