+++ src/components/Transition.vue
@@ -4,9 +4,12 @@
     v-bind:css="true"
     tag="div"
     :move-class="transition.moveClass ? transition.moveClass : 'vt-move'"
-    @leave="setAbsolutePosition"
+    @leave="leave"
+    @before-enter="beforeEnter"
+    @after-enter="afterEnter"
+    @before-leave="beforeLeave"
   >
-    <slot></slot>
+    <slot />
   </transition-group>
 </template>
 
@@ -14,31 +17,56 @@
 export default {
   name: "Transition",
   props: {
-    transition: { type: [String, Object], required: true }
+    transition: { type: [String, Object], required: true },
+    position: { type: String, required: true }
   },
   methods: {
-    setAbsolutePosition(el) {
+    leave(el) {
+      const position = this.position.split("-");
       // https://forum.vuejs.org/t/transition-group-move-class-not-occuring-in-the-array/6381/5
       // these rules ensure the toast stays where it is
       const { height, width, marginBottom } = window.getComputedStyle(el);
       // when the last toast removed the container collapses hence the need for the width subtraction
       el.style.left =
-        el.offsetLeft -
-        (el.parentNode.childNodes.length === 1 ? parseInt(width) : 0) +
+        (el.offsetLeft -
+          (el.parentNode.childNodes.length === 1 ? parseInt(width) : 0)) /
+          2 +
         "px";
-      el.style.top =
-        el.offsetTop - parseInt(height) - parseInt(marginBottom) + "px";
+      el.style.top = el.offsetTop + "px";
+      if (position[0] === "center") {
+        el.style.top =
+          parseInt(el.style.top) -
+          parseInt(height) / 2 -
+          parseInt(marginBottom) / 2 +
+          "px";
+      }
+      if (position[0] === "bottom") {
+        el.style.top =
+          parseInt(el.style.top) -
+          parseInt(height) -
+          parseInt(marginBottom) +
+          "px";
+      }
 
       el.style.position = "absolute";
+    },
+    beforeEnter() {
+      // no delay on making space for notification
+      this.$el.childNodes.forEach(node => delete node.dataset.delayed);
+    },
+    afterEnter(el) {
+      el.classList.add("vt-default-position");
+    },
+    beforeLeave(el) {
+      // this ensures that notifications won't move until the other has been removed
+      for (let i = 0; i < el.parentNode.childNodes.length; i++) {
+        if (el.parentNode.childNodes[i].isSameNode(el)) {
+          continue;
+        }
+        el.parentNode.childNodes[i].dataset.delayed = true;
+      }
+      el.classList.remove("vt-default-position");
     }
   }
 };
 </script>
-
-<style lang="css">
-.vt-move {
-  transition-timing-function: ease-out;
-  transition-property: all;
-  transition-duration: 200ms;
-}
-</style>